Refactored PR
Used kebab naming convention
Put all non pages file into components counterpart
Reworked create team to be more modular
