+++ GameOn-Frontend/app/(tabs)/spaces/_layout.tsx
@@ -2,7 +2,7 @@ import { Stack } from "expo-router";
 import React from "react";
 import Header from "@/components/header/header";
 import { Logo } from "@/components/header/logo";
-import { AddButton } from "@/components/header/addButton";
+import { AddButton } from "@/components/header/add-button";
 import PageTitle from "@/components/header/page-title";
 
 const header = () => (
