+++ GameOn-Frontend/app/_layout.tsx
@@ -9,7 +9,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ClerkProvider, ClerkLoaded } from "@clerk/clerk-expo";
 import { tokenCache } from "@clerk/clerk-expo/token-cache";
 import { useColorScheme } from "@/hooks/use-color-scheme";
-import { FeatureFlagsProvider } from "@/app/(contexts)/featureFlags/FeatureFlagsContext";
+import { FeatureFlagsProvider } from "@/components/feature-flags/feature-flags-context";
 
 const queryClient = new QueryClient();
 export const unstable_settings = { anchor: "(tabs)" };
