+++ GameOn-Frontend/components/teams/teams-styles.ts
